# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AiRunDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "draft_brief": {
            "title": "Draft Brief",
            "type": "boolean"
          },
          "events": {
            "items": {
              "discriminator": {
                "mapping": {
                  "brief": "#/components/schemas/AiBriefEvent",
                  "market": "#/components/schemas/AiMarketEvent",
                  "phase": "#/components/schemas/AiPhaseEvent",
                  "proposal": "#/components/schemas/AiProposalEvent",
                  "queries": "#/components/schemas/AiQueriesEvent",
                  "step": "#/components/schemas/AiStepEvent"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/AiStepEvent"
                },
                {
                  "$ref": "#/components/schemas/AiBriefEvent"
                },
                {
                  "$ref": "#/components/schemas/AiQueriesEvent"
                },
                {
                  "$ref": "#/components/schemas/AiMarketEvent"
                },
                {
                  "$ref": "#/components/schemas/AiProposalEvent"
                },
                {
                  "$ref": "#/components/schemas/AiPhaseEvent"
                }
              ]
            },
            "title": "Events",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "listing": {
            "title": "Listing",
            "type": "string"
          },
          "origin": {
            "enum": [
              "manual",
              "batch"
            ],
            "title": "Origin",
            "type": "string"
          },
          "phase": {
            "enum": [
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Phase",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "id",
          "listing",
          "draft_brief",
          "origin",
          "phase",
          "steps",
          "created_at",
          "finished_at",
          "events"
        ],
        "title": "AiRunDetail",
        "type": "object"
      },
      "AiRunRefusal": {
        "description": "A ``409`` from ``POST /api/ai/runs``: exactly one of the two is set.\n``active_run`` names the run to reattach to; ``reason`` says which\nreadiness rule failed.",
        "properties": {
          "active_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Run"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "AiRunRefusal",
        "type": "object"
      },
      "AiRunSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "draft_brief": {
            "title": "Draft Brief",
            "type": "boolean"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "listing": {
            "title": "Listing",
            "type": "string"
          },
          "origin": {
            "enum": [
              "manual",
              "batch"
            ],
            "title": "Origin",
            "type": "string"
          },
          "phase": {
            "enum": [
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Phase",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "id",
          "listing",
          "draft_brief",
          "origin",
          "phase",
          "steps",
          "created_at",
          "finished_at"
        ],
        "title": "AiRunSummary",
        "type": "object"
      },
...
      "CreateAiRunRequest": {
        "properties": {
          "draft_brief": {
            "default": false,
            "title": "Draft Brief",
            "type": "boolean"
          },
          "listing": {
            "title": "Listing",
            "type": "string"
          }
        },
        "required": [
          "listing"
        ],
        "title": "CreateAiRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanRunSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "plan",
            "title": "Kind",
            "type": "string"
          },
          "listings": {
            "items": {
              "type": "string"
            },
            "title": "Listings",
            "type": "array"
          },
          "phase": {
            "enum": [
              "queued",
              "planning",
              "planned",
              "previewing",
              "ready",
              "failed",
              "cancelled"
            ],
            "title": "Phase",
            "type": "string"
          },
          "reviewed_run_id": {
            "title": "Reviewed Run Id",
            "type": "null"
          },
          "scope": {
            "enum": [
              "listings",
              "workspace"
            ],
            "title": "Scope",
            "type": "string"
          },
          "seen": {
            "title": "Seen",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "scope",
          "listings",
          "seen",
          "created_at",
          "kind",
          "phase"
        ],
        "title": "PlanRunSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "etsy-listings",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ai/runs": {
      "get": {
        "description": "The listing's running or most recent run -- what the editor\nreattaches to after a reload.",
        "operationId": "find_ai_run_api_ai_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "listing",
            "required": true,
            "schema": {
              "title": "Listing",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiRunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find Ai Run",
        "tags": [
          "ai-runs"
        ]
      },
      "post": {
        "description": "Start a run for a saved listing. A ``409`` either names the active run\nto reattach to, or gives the readiness rule that failed.",
        "operationId": "create_ai_run_api_ai_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAiRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiRunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiRunRefusal"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ai Run",
        "tags": [
          "ai-runs"
        ]
      }
    },
    "/api/ai/runs/{run_id}": {
      "delete": {
        "description": "Cancel: the run's provider subprocess tree is killed and research\nstarts no new call. The run ends ``cancelled`` on its own thread\nshortly after; a brief or snapshot already written stays.",
        "operationId": "cancel_ai_run_api_ai_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiRunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Ai Run",
        "tags": [
          "ai-runs"
        ]
      },
      "get": {
        "operationId": "get_ai_run_api_ai_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ai Run",
        "tags": [
          "ai-runs"
        ]
      }
    },
    "/api/ai/runs/{run_id}/events": {
      "get": {
        "operationId": "stream_ai_run_events_api_ai_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Ai Run Events",
        "tags": [
          "ai-runs"
        ]
      }
    },
...
    "/api/runs": {
      "get": {
        "description": "``listing`` narrows to that listing's current run (active, or finished\nand not yet superseded -- decision 7's retention). Omitted, every run this\nprocess still remembers, for a future workspace-wide view.",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "listing",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Listing"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "listings",
                    "workspace"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "discriminator": {
                      "mapping": {
                        "apply": "#/components/schemas/ApplyRunSummary",
                        "plan": "#/components/schemas/PlanRunSummary"
                      },
                      "propertyName": "kind"
                    },
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/PlanRunSummary"
                      },
                      {
                        "$ref": "#/components/schemas/ApplyRunSummary"
                      }
                    ]
                  },
                  "title": "Response List Runs Api Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "A ``409`` names the run already holding one of these listings, so the\ncaller can reattach to it (``GET /api/runs/{active_run}``) instead of\nretrying into the same refusal.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ListingPlanRequest"
                  },
                  {
                    "$ref": "#/components/schemas/WorkspacePlanRequest"
                  },
                  {
                    "$ref": "#/components/schemas/ListingApplyRequest"
                  },
                  {
                    "$ref": "#/components/schemas/WorkspaceApplyRequest"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "apply": "#/components/schemas/ApplyRunSummary",
                      "plan": "#/components/schemas/PlanRunSummary"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PlanRunSummary"
                    },
                    {
                      "$ref": "#/components/schemas/ApplyRunSummary"
                    }
                  ],
                  "title": "Response Create Run Api Runs Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "delete": {
        "description": "Cancel a queued or running **plan**. An ``apply`` is refused\nunconditionally (decision 8: Back leaves, it never cancels one), and so is\na run that has already finished.",
        "operationId": "cancel_run_api_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "apply": "#/components/schemas/ApplyRunSummary",
                      "plan": "#/components/schemas/PlanRunSummary"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PlanRunSummary"
                    },
                    {
                      "$ref": "#/components/schemas/ApplyRunSummary"
                    }
                  ],
                  "title": "Response Cancel Run Api Runs  Run Id  Delete"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "apply": "#/components/schemas/ApplyRunDetail",
                      "plan": "#/components/schemas/PlanRunDetail"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PlanRunDetail"
                    },
                    {
                      "$ref": "#/components/schemas/ApplyRunDetail"
                    }
                  ],
                  "title": "Response Get Run Api Runs  Run Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "description": "``text/event-stream`` over this run's own event buffer -- every event\nalready recorded past ``Last-Event-ID``, then whatever the worker thread\nappends next, until the run reaches a terminal phase.",
        "operationId": "stream_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/seen": {
      "post": {
        "operationId": "mark_seen_api_runs__run_id__seen_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "apply": "#/components/schemas/ApplyRunSummary",
                      "plan": "#/components/schemas/PlanRunSummary"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PlanRunSummary"
                    },
                    {
                      "$ref": "#/components/schemas/ApplyRunSummary"
                    }
                  ],
                  "title": "Response Mark Seen Api Runs  Run Id  Seen Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Seen",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}